Refactoring common parts of DPMM into base class
