+++ CMakeLists.txt
@@ -6,6 +6,6 @@ set(CMAKE_CXX_STANDARD 11)
 add_subdirectory(Rmath)
 
 add_executable(carbondate
-        main.cpp src/WalkerDPMM.cpp src/WalkerDPMM.h src/helpers.cpp src/sort.cpp src/helpers.h src/csv_helpers.cpp src/csv_helpers.h src/DensityOutput.cpp src/DensityOutput.h src/read_data.cpp src/read_data.h src/PosteriorDensityOutput.cpp src/PosteriorDensityOutput.h src/PredictiveDensityOutput.cpp src/PredictiveDensityOutput.h)
+        main.cpp src/WalkerDPMM.cpp src/WalkerDPMM.h src/helpers.cpp src/sort.cpp src/helpers.h src/csv_helpers.cpp src/csv_helpers.h src/DensityOutput.cpp src/DensityOutput.h src/read_data.cpp src/read_data.h src/PosteriorDensityOutput.cpp src/PosteriorDensityOutput.h src/PredictiveDensityOutput.cpp src/PredictiveDensityOutput.h src/DPMM.cpp src/DPMM.h)
 target_include_directories(carbondate PUBLIC Rmath/include)
 target_link_libraries(carbondate PUBLIC Rmath)
